Made button that activates Photo Slider
